requirements.txt
aiohttp==3.7.4.post0
aiomysql==0.0.21
beautifulsoup4==4.9.3
blend-modes==2.1.0
colorama==0.4.4
colory==0.2
discord==1.7.3
discord.py==1.7.3
Google-Images-Search==1.3.8
numpy==1.21.1
Pillow==8.3.1
PyMySQL==0.9.2
PyYAML==6.0.1
python-dateutil==2.8.2
requests==2.21.0
urllib3==1.24.1
virtualenv==16.6.1
mtcnn
opencv-python
numpy